requirements.txt
Django==2.0.7
djangorestframework==3.8.2
pytz==2018.5
drfaddons
drf_contact
django-filter==1.1.0
